extract loadContent() methods into reusable traits
